# public/swagger.json
{
  "components": {
...
    "schemas": {
...
      "EvaNumber": {
        "type": "string",
        "minLength": 7,
        "maxLength": 7
      },
...
      "ParsedProduct": {
        "$ref": "#/components/schemas/CommonProductInfo"
      },
...
      "AllowedHafasProfile": {
        "enum": [
          "db",
          "oebb",
          "bvg",
          "hvv",
          "rmv",
          "sncb",
          "avv",
          "nahsh",
          "insa",
          "anachb",
          "vao",
          "sbb",
          "dbnetz",
          "pkp",
          "dbregio",
          "smartrbl",
          "vbn"
        ],
        "type": "string"
      },
...
      "AuslastungsValue": {
        "description": "1: Gering\n2: Hoch\n3: Sehr Hoch\n4: Ausgebucht",
        "enum": [1, 2, 3, 4],
        "type": "number"
      },
      "Route$Auslastung": {
        "properties": {
          "first": {
            "$ref": "#/components/schemas/AuslastungsValue"
          },
          "second": {
            "$ref": "#/components/schemas/AuslastungsValue"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Route$Stop": {
        "properties": {
          "arrival": {
            "$ref": "#/components/schemas/CommonStopInfo"
          },
          "departure": {
            "$ref": "#/components/schemas/CommonStopInfo"
          },
          "station": {
            "$ref": "#/components/schemas/Station"
          },
          "auslastung": {
            "$ref": "#/components/schemas/Route$Auslastung"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/RemL"
            },
            "type": "array"
          },
          "additional": {
            "type": "boolean"
          },
          "cancelled": {
            "type": "boolean"
          },
          "irisMessages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array"
          }
        },
        "required": ["station"],
        "type": "object",
        "additionalProperties": false
      },
      "ParsedJourneyMatchResponse": {
        "properties": {
          "train": {
            "$ref": "#/components/schemas/ParsedProduct"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Route$Stop"
            },
            "type": "array"
          },
          "jid": {
            "type": "string"
          },
          "firstStop": {
            "$ref": "#/components/schemas/Route$Stop"
          },
          "lastStop": {
            "$ref": "#/components/schemas/Route$Stop"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/RemL"
            },
            "type": "array"
          }
        },
        "required": ["train", "stops", "jid", "firstStop", "lastStop"],
        "type": "object",
        "additionalProperties": false
      },
      "JourneyMatchOptions": {
        "properties": {
          "trainName": {
            "type": "string"
          },
          "initialDepartureDate": {
            "type": "string",
            "format": "date-time"
          },
          "jnyFltrL": {
            "items": {
              "$ref": "#/components/schemas/JourneyFilter"
            },
            "type": "array"
          }
        },
        "required": ["trainName"],
        "type": "object",
        "additionalProperties": false
      },
...
      "HafasStation": {
        "properties": {
          "title": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ParsedProduct"
            },
            "type": "array"
          },
          "coordinates": {
            "$ref": "#/components/schemas/Coordinates"
          }
        },
        "required": ["title", "id", "coordinates"],
        "type": "object",
        "additionalProperties": false
      },
...
      "ParsedJourneyGeoPosResponse": {
        "items": {
          "$ref": "#/components/schemas/SingleParsedJourneyGeoPos"
        },
        "type": "array"
      },
...
      "JourneyGeoPosOptions": {
        "$ref": "#/components/schemas/JourneyGeoPosRequestOptions"
      },
...
      "ParsedJourneyDetails": {
        "properties": {
          "train": {
            "$ref": "#/components/schemas/ParsedProduct"
          },
          "auslastung": {
            "$ref": "#/components/schemas/Route$Auslastung"
          },
          "jid": {
            "type": "string"
          },
          "firstStop": {
            "$ref": "#/components/schemas/Route$ValidDepartureStop"
          },
          "lastStop": {
            "$ref": "#/components/schemas/Route$ValidArrivalStop"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Route$Stop"
            },
            "type": "array"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/RemL"
            },
            "type": "array"
          }
        },
        "required": ["train", "jid", "firstStop", "lastStop", "stops"],
        "type": "object",
        "additionalProperties": false
      },
      "ArrivalStationBoardEntry": {
        "properties": {
          "train": {
            "$ref": "#/components/schemas/ParsedProduct"
          },
          "cancelled": {
            "type": "boolean"
          },
          "finalDestination": {
            "type": "string"
          },
          "jid": {
            "type": "string"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Route$Stop"
            },
            "type": "array"
          },
          "currentStation": {
            "$ref": "#/components/schemas/HafasStation"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/RemL"
            },
            "type": "array"
          },
          "arrival": {
            "$ref": "#/components/schemas/CommonStopInfo"
          }
        },
        "required": [
          "train",
          "finalDestination",
          "jid",
          "currentStation",
          "arrival"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "DepartureStationBoardEntry": {
        "properties": {
          "train": {
            "$ref": "#/components/schemas/ParsedProduct"
          },
          "cancelled": {
            "type": "boolean"
          },
          "finalDestination": {
            "type": "string"
          },
          "jid": {
            "type": "string"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Route$Stop"
            },
            "type": "array"
          },
          "currentStation": {
            "$ref": "#/components/schemas/HafasStation"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/RemL"
            },
            "type": "array"
          },
          "departure": {
            "$ref": "#/components/schemas/CommonStopInfo"
          }
        },
        "required": [
          "train",
          "finalDestination",
          "jid",
          "currentStation",
          "departure"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "SecL": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/SecLJNY"
          },
          {
            "$ref": "#/components/schemas/SecLWALK"
          },
          {
            "$ref": "#/components/schemas/SecLKISS"
          }
        ]
      },
...
      "ParsedSearchOnTripResponse": {
        "properties": {
          "cancelled": {
            "type": "boolean"
          },
          "changeDuration": {
            "type": "number",
            "format": "double"
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "finalDestination": {
            "type": "string"
          },
          "jid": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/ProdL"
          },
          "raw": {
            "$ref": "#/components/schemas/SecL"
          },
          "segmentDestination": {
            "$ref": "#/components/schemas/Station"
          },
          "segmentStart": {
            "$ref": "#/components/schemas/Station"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Route$Stop"
            },
            "type": "array"
          },
          "train": {
            "$ref": "#/components/schemas/ParsedProduct"
          },
          "auslastung": {
            "$ref": "#/components/schemas/Route$Auslastung"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/RemL"
            },
            "type": "array"
          },
          "tarifSet": {
            "items": {
              "$ref": "#/components/schemas/Route$TarifFareSet"
            },
            "type": "array"
          },
          "plannedSequence": {
            "$ref": "#/components/schemas/PlannedSequence"
          },
          "type": {
            "type": "string",
            "enum": ["JNY"],
            "nullable": false
          },
          "arrival": {
            "$ref": "#/components/schemas/CommonStopInfo"
          },
          "departure": {
            "$ref": "#/components/schemas/CommonStopInfo"
          },
          "wings": {
            "items": {
              "$ref": "#/components/schemas/Route$Journey"
            },
            "type": "array"
          },
          "currentStop": {
            "$ref": "#/components/schemas/Route$Stop"
          }
        },
        "required": [
          "finalDestination",
          "jid",
          "segmentDestination",
          "segmentStart",
          "stops",
          "train",
          "type",
          "arrival",
          "departure"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "SingleRoute": {
        "properties": {
          "arrival": {
            "$ref": "#/components/schemas/CommonStopInfo"
          },
          "departure": {
            "$ref": "#/components/schemas/CommonStopInfo"
          },
          "isRideable": {
            "type": "boolean"
          },
          "checksum": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "changes": {
            "type": "integer",
            "format": "int32"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/Route$JourneySegment"
            },
            "type": "array"
          },
          "segmentTypes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tarifSet": {
            "items": {
              "$ref": "#/components/schemas/Route$TarifFareSet"
            },
            "type": "array"
          },
          "raw": {
            "$ref": "#/components/schemas/OutConL"
          }
        },
        "required": [
          "arrival",
          "departure",
          "isRideable",
          "checksum",
          "cid",
          "date",
          "duration",
          "changes",
          "segments",
          "segmentTypes"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "AllowedSotMode": {
        "type": "string",
        "enum": ["JI", "RC"]
      },
      "SearchOnTripBody": {
        "properties": {
          "sotMode": {
            "$ref": "#/components/schemas/AllowedSotMode"
          },
          "id": {
            "type": "string"
          }
        },
        "required": ["sotMode", "id"],
        "type": "object",
        "additionalProperties": false
      },
      "RoutingResult": {
        "properties": {
          "routes": {
            "items": {
              "$ref": "#/components/schemas/SingleRoute"
            },
            "type": "array"
          },
          "context": {
            "properties": {
              "later": {
                "type": "string"
              },
              "earlier": {
                "type": "string"
              }
            },
            "required": ["later", "earlier"],
            "type": "object"
          }
        },
        "required": ["routes", "context"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SBBStation": {
        "properties": {
          "title": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/SBBCoordinates"
          }
        },
        "required": ["title", "id", "location"],
        "type": "object",
        "additionalProperties": false
      },
      "RoutingOptions": {
        "properties": {
          "start": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["start", "destination"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Bahnhofs Abfahrten",
    "version": "1.0.0",
    "contact": {
      "name": "marudor"
    }
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/hafas/v2/arrivalStationBoard": {
      "get": {
        "operationId": "Arrival Station Board v2",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArrivalStationBoardEntry"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "description": "evaNumber",
            "in": "query",
            "name": "station",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EvaNumber"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ]
      }
    },
    "/hafas/v2/auslastung/{start}/{destination}/{trainNumber}/{time}": {
      "get": {
        "operationId": "Auslastung v2",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route$Auslastung"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "start",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "destination",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "trainNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "time",
            "required": true,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/hafas/v2/departureStationBoard": {
      "get": {
        "operationId": "Departure Station Board v2",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DepartureStationBoardEntry"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "station",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EvaNumber"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EvaNumber"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ]
      }
    },
    "/hafas/v2/details/{trainName}": {
      "get": {
        "operationId": "Details v2",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedSearchOnTripResponse"
                }
              }
            }
          },
          "404": {
            "description": "Train not found"
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "trainName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stop",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "EvaNumber of a stop of your train",
            "in": "query",
            "name": "station",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EvaNumber"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ]
      }
    },
...
    "/hafas/v1/geoStation": {
      "get": {
        "operationId": "Geo Station v1",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HafasStation"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "maxDist",
            "required": false,
            "schema": {
              "default": 1000,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ]
      }
    },
...
    "/hafas/v1/journeyGeoPos": {
      "post": {
        "operationId": "Journey Geo Position v1",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedJourneyGeoPosResponse"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyGeoPosOptions"
              }
            }
          }
        }
      }
    },
...
    "/hafas/v2/journeyMatch": {
      "post": {
        "operationId": "Journey Match v2",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParsedJourneyMatchResponse"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyMatchOptions"
              }
            }
          }
        }
      }
    },
...
    "/hafas/v2/journeyDetails": {
      "get": {
        "operationId": "JourneyDetails v2",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedJourneyDetails"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "jid",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ]
      }
    },
...
    "/hafas/v1/positionForTrain/{trainName}": {
      "get": {
        "operationId": "Position for Train v1",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedJourneyGeoPosResponse"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "trainName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ]
      }
    },
    "/sbb/experimental/routing": {
      "post": {
        "operationId": "Routing",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": ["SBB Experimental"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutingOptions"
              }
            }
          }
        }
      }
    },
    "/hafas/v2/searchOnTrip": {
      "post": {
        "operationId": "Search on Trip v2",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleRoute"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchOnTripBody"
              }
            }
          }
        }
      }
    },
...
    "/sbb/experimental/station/{searchTerm}": {
      "get": {
        "operationId": "Station",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SBBStation"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["SBB Experimental"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "searchTerm",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/hafas/v3/tripSearch": {
      "post": {
        "operationId": "TripSearch v3",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingResult"
                }
              }
            }
          }
        },
        "tags": ["HAFAS"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllowedHafasProfile"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripSearchOptionsV3"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://marudor.de/api"
    }
  ]
}